+++ R/AFMdata.R
@@ -277,6 +277,8 @@ AFM.raster <- function(obj,no=1) {
 }
 
 #' Graph of AFMdata object
+#' 
+#' By default, trims 1 percent of the outliers in height data
 #'
 #' @param x AFMdata object
 #' @param no channel number of the image
@@ -297,7 +299,7 @@ AFM.raster <- function(obj,no=1) {
 #' d = AFM.import(AFM.getSampleImages(type='ibw')[1])
 #' plot(d, graphType=2)
 #' @export
-plot.AFMdata <- function(x, no=1, mpt=NA, graphType=1, trimPeaks=0, addLines=FALSE, redBlue = FALSE, verbose=FALSE, quiet=FALSE, ...) {
+plot.AFMdata <- function(x, no=1, mpt=NA, graphType=1, trimPeaks=0.01, addLines=FALSE, redBlue = FALSE, verbose=FALSE, quiet=FALSE, ...) {
   if (no>length(x@channel)) stop("imageNo out of bounds.")
   if (!quiet) cat("Graphing:",x@channel[no])
   if (verbose) print(paste("History:",x@history))